Add repeat qdel check to create_and_destroy Retry some changes to atoms subsystem

+++ code/modules/unit_tests/create_and_destroy.dm
@@ -59,7 +59,8 @@ GLOBAL_VAR_INIT(running_create_and_destroy, FALSE)
 		var/list/to_del = spawn_at.contents - cached_contents
 		if(length(to_del))
 			for(var/atom/to_kill in to_del)
-				qdel(to_kill)
+				if(!QDELETED(to_kill))
+					qdel(to_kill)
 
 	GLOB.running_create_and_destroy = FALSE
 	//Hell code, we're bound to have ended the round somehow so let's stop if from ending while we work
